Environment tests
Refactoring tests to provide more readability and organization.
